configs: r8a77965: Fix XTAL clock for Salvator-XS board
This patch is changes to default configurations as Salvator-XS board.
At the case of use Salvator-X board, need modify below.

modify at "include/configs/r8a77965_salvator-x_defconfig".
When "BOARD_SALVATOR_XS" is defined, it is for setting Salvator-XS.
When "BOARD_SALVATOR_XS" is not defined, it is for setting Salvator-X.

Signed-off-by: Hiroyuki Yokoyama <hiroyuki.yokoyama.vx@renesas.com>