ARM: rmobile: Introduce RCAR_HAS_SECURE_BASE config
Introduce new config to indicate that U-Boot start address belongs to
secure region. In that case, adjust the access range and the mapped area
accordingly. This is especially useful if we want to be able to run a
minimal PSCI handling in U-Boot, where the PSCI sections are placed
together with the boot loader in secure region.

Signed-off-by: Hai Pham <hai.pham.ud@renesas.com>