mmc: Introduce mmc_abort_tuning()
Based on Linux commit e711f03091097 ("mmc: mmc: Introduce
mmc_abort_tuning()")

If a tuning command times out, the card could still be processing it,
which will cause problems for recovery. The eMMC specification says that
CMD12 can be used to stop CMD21, so add a function that does that.

Signed-off-by: Hai Pham <hai.pham.ud@renesas.com>
Reviewed-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>