net: rswtich: disable GWCA when removing driver
We need to disable GWCA for two reasons:

1. There may be packets coming to the device when U-Boot already
transferred control to Linux, but Linux didn't had chance to
re-configure GWCA. This may lead to memory corruption, as chains are
already active.

2. Linux will not be able to switch GWCA to DISABLED state, because
transition from OPERATION to DISABLED requires that all buffer
pointers are released and all transactions are finished. But there are
no one who can handle those operations, as U-Boot already exited and
Linux is still trying to initialize R-Switch driver.

This fixes issue with DomD unable to initialize network in the
parallel mode.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>